Add "home" and "end" keyboard controls on unix, Ctrl-C to clear, then to exit
Windows will also work if somebody figures out the keycodes.  Currently I've left them as -1

The first Ctrl-C clears the current input (bash style), the second exits the program.

This is similar to behaviour in Python's shell, where Ctrl-D only exits if the line is
already empty.  It also matches the usage of Bash, where Ctrl-C clears the current line,
or exits the current process.

Hopefully it's the best of both worlds, but open for discussion