add -codi switch so ihx is compatible w/ codi.vim
Codi is an interactive scratchpad for hackers.
https://github.com/metakirby5/codi.vim

Because of ihx way of continuously redrawing the prompted line it
couldn't work with codi. The hereafter modifications enable
(with haxelib run ihx -codi) a very simple alternate interpreter
that will work with codi.

Here is my configuration in .vimrc:
-------------------------------------
let g:codi#interpreters = {
    \ 'haxe': {
        \ 'bin': ['/home/xx/bin/ihx', '-codi'],
        \ 'prompt': '>> ',
        \ 'quitcmd': 'exit',
    \ },
\ }